Implement compression configuration

+++ src/deploy.sh
@@ -1,55 +1,58 @@
 #!/bin/sh
 
+
+# $1 = source , $2 = package , $3 = name
 package()
 {
-  unset clean_needed
   src="$1"
   pkg="$2"
-  echo "Packaging $(getname "$src"): $(du -sh "$src" | awk '{print $1}')iB"
 
-  tmpdir="/tmp/zpkg_$(random_string 5)"
-  mkdir -p "$tmpdir"
-  if [ ! -d "$src/ROOT" ] && [ ! -d "$src/HOME" ] && [ ! -f "$src/DEPS" ]
+  unset tmpdir
+  if [ ! -d "$src/ROOT" ] && [ ! -d "$src/HOME" ] && [ ! -f "$src/DEPS" ] && [ ! -f "$src/DESC" ]
   then
-    mkdir -p "$tmpdir/package"
-    cp -r "$src" "$tmpdir/package/ROOT"
-  else
-    cp -r "$src" "$tmpdir/package"
+    tmpdir="/tmp/zpkg_$(random_string 5)"
+    mkdir -p "$tmpdir"
+    cp -r "$src" "$tmpdir/ROOT"
+    src="$tmpdir"
   fi
   (
-  cd "$tmpdir/package"
+  cd "$src"
   unset list
   [ -f DEPS ] && list=DEPS
   [ -f DESC ] && list="$list DESC"
   [ -d HOME ] && list="$list HOME"
   [ -d ROOT ] && list="$list ROOT"
-  if which pv >/dev/null 2>&1
-  then
-    tar -cf - --owner=0 --group=0 -P $list | pv -s "$(du -sb . | awk '{print $1}')" | xz > "../$pkg"
-  else
-    tar -cvJf - --owner=0 --group=0 $list > "../$pkg"
-  fi
+  size=$(du -sb $list | awk '{print $1}' | paste -sd+ | bc)
+  echo "Packaging $(basename "$pkg"): $(echo "$size" | numfmt --to=iec-i)B"
+  cc=$compress
+  [ $size -gt 1048576 ] && cc=$pcompress
+  tar -cf - --owner=0 --group=0 $list | $cc $comparg > "$pkg"
   )
-  mv "$tmpdir/$pkg" ./
-  rm -rd "$tmpdir"
+  [ -n "$tmpdir" ] && rm -rd "$tmpdir"
+  return 0
 }
 
+# $1 = file
 deploy_package()
 {
-  echo "Deploying $1: $(du -sh "$1" | awk '{print $1}')iB"
+  echo "Deploying $(basename "$1"): $(du -sh "$1" | awk '{print $1}')iB"
   scp "$1" $SSH_ADDRESS:~/'$(grep "PKG_PATH=" .config | cut -d"=" -f2-)'
 }
 
 deploy_folder()
 {
-  archive="$(getname "$1").tar.xz"
-  if [ -n "$(echo "$1" | grep '\.tar\.xz$' )" ]
+  if [ -f "$1" ] && echo "$1" | grep -q '\.tar\.'"$extension\$" # file and valid extension
   then
-    deploy_package "$1" || return 1
+    $pcompress -dc >/dev/null 2>&1 | tar -tf - >/dev/null 2>&1|| { echo "File '$1' is not a valid archive" && return 1; }
+    deploy_package "$1" "$1" || return $?
+  elif [ -d "$1" ]
+  then
+    archive="$(getname "$1").tar.$extension"
+    package "$1" "/tmp/$archive" || return $?
+    deploy_package "/tmp/$archive" || return $?
+    rm "/tmp/$archive"
   else
-    package "$1" "$archive" || return 1
-    deploy_package "$archive" || return 1
-    rm "$archive" 2> /dev/null
+    echo "Target '$1' doesn't exist"
   fi
 }
